+++ CHANGELOG.md
@@ -13,18 +13,23 @@ Types of changes
 
 ## Unreleased
 
+## 2024-10-08, v0.3.4
+
 ### Added
 
-- leap second difference between two datetimes: method 'diffLeap', returns Duration type.
+- leap second difference between two datetimes: method 'diffLeap', returns Duration type
+- 'replace' method to safely change fields of a datetime
+- 'toFields' method for datetime
 
 ### Changed
 
-- tzdb version info now part of tzdata.zig, since this is specific to the embedded version.
+- tzdb version info now part of tzdata.zig, since this is specific to the embedded version
 - update info on how to update tzdata for development
+- isocalendar: 'year' field is now called 'isoyear'
 
 ### Removed
 
-- gen_tzdb_version script. This is now part of the tzdb_update script.
+- gen_tzdb_version script; this is now part of the tzdb_update script
 
 ## 2024-10-04, v0.3.3